pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "monarchmoneycommunity>=1.3.0",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tenacity>=9.1.4",
    "textual>=8.0.0",
]

[tool.ruff]
line-length = 200
